moved customer state into container component
