+++ src/components/pages/CustomerProfile/CustomerProfileContainer.js
@@ -8,6 +8,9 @@ const CustomerProfileContainer = props => {
   const [contactModalVisible, setContactModalVisible] = useState(false);
   const [profileModalVisible, setProfileModalVisible] = useState(false);
   const [aboutModalVisible, setAboutModalVisible] = useState(false);
+  const [profileInfo, setProfileInfo] = useState({});
+  const [message, setMessage] = useState('');
+
   const { getCustomerInfo } = props;
   const customerId = localStorage.getItem('customerId');
 
@@ -15,8 +18,32 @@ const CustomerProfileContainer = props => {
     getCustomerInfo(customerId);
   }, [getCustomerInfo, customerId]);
 
+  const handleChange = e => {
+    setProfileInfo({
+      ...profileInfo,
+      [e.target.name]: e.target.value,
+    });
+  };
+
+  const validateForm = () => {
+    for (let input in profileInfo) {
+      if (input !== 'photo_url') {
+        let value = profileInfo[input];
+        value = value.replace(/^\s+/, '').replace(/\s+$/, '');
+        if (value === '') {
+          return false;
+        }
+      }
+    }
+    return true;
+  };
+
   const updateProfile = data => {
-    props.updateCustomer(data, customerId);
+    if (validateForm()) {
+      props.updateCustomer(data, customerId);
+    } else {
+      setMessage('This field is required');
+    }
   };
 
   const showContactModal = () => {
@@ -59,6 +86,12 @@ const CustomerProfileContainer = props => {
         updateProfile={updateProfile}
         error={props.error}
         status={props.status}
+        profileInfo={profileInfo}
+        setProfileInfo={setProfileInfo}
+        message={message}
+        setMessage={setMessage}
+        handleChange={handleChange}
+        validateForm={validateForm}
       />
     );
   } else if (props.isFetching === true) {
